chore: finalize refactoring
